+++ src/tightdb/objc/TDBQuery.h
@@ -38,15 +38,21 @@
 -(NSUInteger)countRows;
 -(NSUInteger)removeRows;
 
--(int64_t)sumIntColumnWithIndex:(NSUInteger)colIndex;
--(double)sumFloatColumnWithIndex:(NSUInteger)colIndex;
--(double)sumDoubleColumnWithIndex:(NSUInteger)colIndex;
--(int64_t)maxIntInColumnWithIndex:(NSUInteger)colIndex;
--(float)maxFloatInColumnWithIndex:(NSUInteger)colIndex;
--(double)maxDoubleInColumnWithIndex:(NSUInteger)colIndex;
+// Aggregates
 -(int64_t)minIntInColumnWithIndex:(NSUInteger)colIndex;
 -(float)minFloatInColumnWithIndex:(NSUInteger)colIndex;
 -(double)minDoubleInColumnWithIndex:(NSUInteger)colIndex;
+-(NSDate *)minDateInColumnWithIndex:(NSUInteger)colIndex;
+
+-(int64_t)maxIntInColumnWithIndex:(NSUInteger)colIndex;
+-(float)maxFloatInColumnWithIndex:(NSUInteger)colIndex;
+-(double)maxDoubleInColumnWithIndex:(NSUInteger)colIndex;
+-(NSDate *)maxDateInColumnWithIndex:(NSUInteger)colIndex;
+
+-(int64_t)sumIntColumnWithIndex:(NSUInteger)colIndex;
+-(double)sumFloatColumnWithIndex:(NSUInteger)colIndex;
+-(double)sumDoubleColumnWithIndex:(NSUInteger)colIndex;
+
 -(double)avgIntColumnWithIndex:(NSUInteger)colIndex;
 -(double)avgFloatColumnWithIndex:(NSUInteger)colIndex;
 -(double)avgDoubleColumnWithIndex:(NSUInteger)colIndex;