+++ doc/ref/examples/ex_objc_query_dynamic_intro.m
@@ -28,8 +28,8 @@ void ex_objc_query_dynamic_intro()
         TDBView *view = [q findAllRows];
 
         /* Print the names. */
-        for (TDBRow *c in view) {
-            NSLog(@"name: %@",c[NAME]);
+        for (TDBRow *row in view) {
+            NSLog(@"name: %@",row[NAME]);
         }
     }
 }