+++ src/tightdb/objc/TightdbFast.h
@@ -18,13 +18,6 @@
  *
  **************************************************************************/
 
-#import <tightdb/objc/TDBTable.h>
-#import <tightdb/objc/TDBView.h>
-#import <tightdb/objc/TDBQuery.h>
-#import <tightdb/objc/TDBRow.h>
-#import <tightdb/objc/TDBContext.h>
-#import <tightdb/objc/TDBVersion.h>
+#import <tightdb/objc/Tightdb.h>
+#import <tightdb/objc/TDBRowFast.h>
 
-#import <tightdb/objc/PrivateHelperMacros.h>
-#import <tightdb/objc/PrivateTableMacros.h>
-#import <tightdb/objc/PrivateTDB.h>