+++ src/tightdb/objc/util.mm
@@ -24,58 +24,14 @@
 #include <tightdb/binary_data.hpp>
 #include <tightdb/string_data.hpp>
 
+#include <vector>
+
 #import "TDBTable.h"
 #import "TDBTable_noinst.h"
 #import "util_noinst.hpp"
 #import "NSData+TDBGetBinaryData.h"
 
 using namespace tightdb;
-
-inline bool nsnumber_is_like_bool(NSObject *obj)
-{
-    const char* data_type = [(NSNumber *)obj objCType];
-    /* @encode(BOOL) is 'B' on iOS 64 and 'c'
-     objcType is always 'c'. Therefore compare to "c".
-     */
-    return data_type[0] == 'c';
-}
-
-inline bool nsnumber_is_like_integer(NSObject *obj)
-{
-    const char* data_type = [(NSNumber *)obj objCType];
-    return (strcmp(data_type, @encode(int)) == 0 ||
-            strcmp(data_type, @encode(long)) ==  0 ||
-            strcmp(data_type, @encode(long long)) == 0 ||
-            strcmp(data_type, @encode(unsigned int)) == 0 ||
-            strcmp(data_type, @encode(unsigned long)) == 0 ||
-            strcmp(data_type, @encode(unsigned long long)) == 0);
-}
-
-inline bool nsnumber_is_like_float(NSObject *obj)
-{
-    const char* data_type = [(NSNumber *)obj objCType];
-    return (strcmp(data_type, @encode(float)) == 0 ||
-            strcmp(data_type, @encode(int)) == 0 ||
-            strcmp(data_type, @encode(long)) ==  0 ||
-            strcmp(data_type, @encode(long long)) == 0 ||
-            strcmp(data_type, @encode(unsigned int)) == 0 ||
-            strcmp(data_type, @encode(unsigned long)) == 0 ||
-            strcmp(data_type, @encode(unsigned long long)) == 0);
-}
-
-inline bool nsnumber_is_like_double(NSObject *obj)
-{
-    const char* data_type = [(NSNumber *)obj objCType];
-    return (strcmp(data_type, @encode(double)) == 0 ||
-            strcmp(data_type, @encode(float)) == 0 ||
-            strcmp(data_type, @encode(int)) == 0 ||
-            strcmp(data_type, @encode(long)) ==  0 ||
-            strcmp(data_type, @encode(long long)) == 0 ||
-            strcmp(data_type, @encode(unsigned int)) == 0 ||
-            strcmp(data_type, @encode(unsigned long)) == 0 ||
-            strcmp(data_type, @encode(unsigned long long)) == 0);
-}
-
 
 void to_mixed(id value, Mixed& m)
 {
@@ -112,6 +68,93 @@ void to_mixed(id value, Mixed& m)
     }
     if ([value isKindOfClass:[TDBTable class]])
         m = Mixed(Mixed::subtable_tag());
+}
+
+
+NSObject* get_cell(size_t col_ndx, size_t row_ndx, Table& table)
+{
+    DataType type = table.get_column_type(col_ndx);
+    switch (type) {
+        case type_String: {
+            NSString *s = [NSString stringWithUTF8String:table.get_string(col_ndx, row_ndx).data()];
+            return s;
+        }
+        case type_Int: {
+            NSNumber *n = [NSNumber numberWithLongLong:table.get_int(col_ndx, row_ndx)];
+            return n;
+        }
+        case type_Float: {
+            NSNumber *n = [NSNumber numberWithFloat:table.get_float(col_ndx, row_ndx)];
+            return n;
+        }
+        case type_Double: {
+            NSNumber *n = [NSNumber numberWithDouble:table.get_double(col_ndx, row_ndx)];
+            return n;
+        }
+        case type_Bool: {
+            NSNumber *n = [NSNumber numberWithBool:table.get_bool(col_ndx, row_ndx)];
+            return n;
+        }
+        case type_Binary: {
+            BinaryData bd = table.get_binary(col_ndx, row_ndx);
+            NSData *d = [NSData dataWithBytes:bd.data() length:bd.size()];
+            return d;
+        }
+        case type_Table: {
+            TDBTable *t = [[TDBTable alloc] init];
+            TableRef table_ref = table.get_subtable(col_ndx, row_ndx);
+            [t setNativeTable:table_ref.get()];
+            return t;
+        }
+        case type_DateTime: {
+            NSDate *d = [NSDate dateWithTimeIntervalSince1970:table.get_datetime(col_ndx, row_ndx).get_datetime()];
+            return d;
+        }
+        case type_Mixed: {
+            Mixed m = table.get_mixed(col_ndx, row_ndx);
+            switch (m.get_type()) {
+                case type_String: {
+                    NSString *s = [NSString stringWithUTF8String:m.get_string().data()];
+                    return s;
+                }
+                case type_Int: {
+                    NSNumber *n = [NSNumber numberWithLongLong:m.get_int()];
+                    return n;
+                }
+                case type_Float: {
+                    NSNumber *n = [NSNumber numberWithFloat:m.get_float()];
+                    return n;
+                }
+                case type_Double: {
+                    NSNumber *n = [NSNumber numberWithDouble:m.get_double()];
+                    return n;
+                }
+                case type_Bool: {
+                    NSNumber *n = [NSNumber numberWithBool:m.get_bool()];
+                    return n;
+                }
+                case type_Binary: {
+                    BinaryData bd = m.get_binary();
+                    NSData *d = [NSData dataWithBytes:bd.data() length:bd.size()];
+                    return d;
+                }
+                case type_Table: {
+                    TDBTable *t = [[TDBTable alloc] init];
+                    TableRef table_ref = table.get_subtable(col_ndx, row_ndx);
+                    [t setNativeTable:table_ref.get()];
+                    return t;
+                }
+                case type_DateTime: {
+                    NSDate *d = [NSDate dateWithTimeIntervalSince1970:m.get_datetime().get_datetime()];
+                    return d;
+                }
+                case type_Mixed:
+                    TIGHTDB_ASSERT(false);
+            }
+        }
+    }
+    TIGHTDB_ASSERT(false);
+    return nil; // make clang happy
 }
 
 
@@ -616,3 +659,101 @@ void set_row_with_labels(size_t row_ndx, Table& table, NSDictionary *data)
         set_cell(col_ndx, row_ndx, table, value);
     }
 }
+
+BOOL set_columns_aux(TableRef& parent, std::vector<size_t> path, NSArray *schema)
+{
+    size_t list_count = [schema count];
+    if (list_count % 2 != 0) {
+        //Error: "Invalid number of entries in schema"
+        return NO;
+    }
+
+    for (size_t i = 0; i < list_count; i += 2) {
+        NSString *key   = [schema objectAtIndex: i];
+        id        value = [schema objectAtIndex: i+1];
+
+        if (![key isKindOfClass:[NSString class]]) {
+            // Error: "Column name must be a string"
+            return NO;
+        }
+
+        try {
+            DataType type;
+            BOOL need_index = false;
+            if ([value isKindOfClass:[NSString class]]) {
+                if ([value isEqualToString:@"string"]) {
+                    type = type_String;
+                }
+                else if ([value isEqualToString:@"string:indexed"]) {
+                    type = type_String;
+                    need_index = YES;
+                }
+                else if ([value isEqualToString:@"binary"]) {
+                    type = type_Binary;
+                }
+                else if ([value isEqualToString:@"int"]) {
+                    type = type_Int;
+                }
+                else if ([value isEqualToString:@"float"]) {
+                    type = type_Float;
+                }
+                else if ([value isEqualToString:@"double"]) {
+                    type = type_Double;
+                }
+                else if ([value isEqualToString:@"bool"]) {
+                    type = type_Bool;
+                }
+                else if ([value isEqualToString:@"date"]) {
+                    type = type_DateTime;
+                }
+                else if ([value isEqualToString:@"mixed"]) {
+                    type = type_Mixed;
+                }
+                else {
+                    // Error: "Invalid column type. Can be \"bool\", \"int\", \"date\", \"string\", \"binary\" or \"mixed\"."
+                    return NO;
+                }
+            }
+            else if ([value isKindOfClass:[NSArray class]]) {
+                type = type_Table;
+            }
+            else {
+                // Error:  "Invalid column type. Can be \"bool\", \"int\", \"date\", \"string\", \"binary\", \"mixed\" or \"[]\"."
+                return NO;
+            }
+
+            size_t column_ndx;
+            StringData column_name([(NSString *)key UTF8String]);
+            if (path.size() > 0) {
+                column_ndx = (*parent).add_subcolumn(path, type, column_name);
+            }
+            else {
+                column_ndx = (*parent).add_column(type, column_name);
+            }
+
+            if (need_index) {
+                (*parent).set_index(column_ndx);
+            }
+
+            if (type == type_Table) {
+                path.push_back(column_ndx);
+                if (!set_columns_aux(parent, path, value)) {
+                    return false;
+                }
+                path.pop_back();
+            }
+        }
+        catch (...) {
+            // Error: "Exception during schema creation"
+            return NO;
+        }
+    }
+    return YES;
+}
+
+BOOL set_columns(TableRef& parent, NSArray *schema)
+{
+    std::vector<size_t> v;
+    return set_columns_aux(parent, v, schema);
+}
+