+++ release_notes.md
@@ -38,7 +38,7 @@ The Objective-C API has been updated and your code will break!
 
 
 ### Enhancements:
-* 
+* Added `minDateInColumnWithIndex` and `maxDateInColumnWithIndex` to `TDBQuery`.
 
 
 =============================================================