+++ release_notes.md
@@ -35,7 +35,7 @@ The Objective-C API has been updated and your code will break!
 * `columnTypeOfColumn:` renamed to `columnTypeOfColumnWithIndex` in `TDBTable`.
 * `columnNameOfColumn:` renamed to `nameOfColumnWithIndex:` in `TDBTable`.
 * `addColumnWithName:andType:` renamed to `addColumnWithName:type:` in `TDBDescriptor`.
-* Fast getters and setters moved from `TDBRow` to `TDBRowFast`.
+* Fast getters and setters moved from `TDBRow.h` to `TDBRowFast.h`.
 
 
 ### Enhancements: