+++ release_notes.md
@@ -19,10 +19,13 @@ The Objective-C API has been updated and your code will break!
 * `???`
 
 =============================================================
-0.5.0 Release notes (yyyy—MM-dd)
+0.5.0 Release notes (2014-04-02)
 Objective-C
 -----------
 The Objective-C API has been updated and your code will break!
+Of notable changes a fast interface has been added. 
+This interface includes specific methods to get and set values into Tightdb.
+To use these methods import <Tightdb/TightdbFast.h>.
 
 ### Bugfixes:
 
@@ -35,10 +38,15 @@ The Objective-C API has been updated and your code will break!
 * `columnTypeOfColumn:` renamed to `columnTypeOfColumnWithIndex` in `TDBTable`.
 * `columnNameOfColumn:` renamed to `nameOfColumnWithIndex:` in `TDBTable`.
 * `addColumnWithName:andType:` renamed to `addColumnWithName:type:` in `TDBDescriptor`.
+* Fast getters and setters moved from `TDBRow.h` to `TDBRowFast.h`.
 
+### Enhancements:
 
 ### Enhancements:
 * Added `minDateInColumnWithIndex` and `maxDateInColumnWithIndex` to `TDBQuery`.
+* Transactions can now be started directly on named tables.
+* You can create dynamic tables with initial schema.
+* `TDBTable` and `TDBView` now have a shared protocol so they can easier be used interchangeably.
 
 
 =============================================================